requirements.txt
asgiref==3.7.2
Django==5.0.5
django-appconf==1.0.5
django-imagekit==5.0.0
pilkit==3.0
Pillow==10.3.0
sqlparse==0.4.4
tzdata==2023.3
